backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.19.0
python-socketio==5.10.0
python-multipart==0.0.6
gradio==4.7.1
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
pytest==7.4.3